@@ -4,17 +4,17 @@
   "private": true,
   "dependencies": {
     "axios": "^0.19.0",
-    "d3-array": "^2.0.2",
+    "d3-array": "^2.3.3",
     "d3-axis": "^1.0.12",
     "d3-ease": "^1.0.5",
-    "d3-scale": "^2.1.2",
+    "d3-scale": "^3.2.0",
     "d3-selection": "^1.3.2",
     "d3-shape": "^1.2.2",
     "d3-transition": "^1.1.3",
     "enzyme": "^3.9.0",
-    "react": "^16.10.2",
+    "react": "^16.11.0",
     "react-datepicker": "^2.9.6",
-    "react-dom": "^16.10.2",
+    "react-dom": "^16.11.0",
     "react-redux": "^7.1.1",
     "react-router-dom": "^5.1.2",
     "react-scripts": "^3.2.0",
@@ -30,15 +30,15 @@
     "build-storybook": "build-storybook -s public"
   },
   "devDependencies": {
-    "@storybook/addon-actions": "^5.1.11",
-    "@storybook/addon-links": "^5.1.11",
-    "@storybook/addon-viewport": "^5.1.11",
-    "@storybook/addons": "^5.1.11",
-    "@storybook/react": "^5.1.11",
+    "@storybook/addon-actions": "^5.2.5",
+    "@storybook/addon-links": "^5.2.5",
+    "@storybook/addon-viewport": "^5.2.5",
+    "@storybook/addons": "^5.2.5",
+    "@storybook/react": "^5.2.5",
     "enzyme-adapter-react-16": "^1.11.2",
-    "node-sass": "^4.12.0",
-    "react-styleguidist": "^9.1.16",
-    "react-test-renderer": "^16.10.2",
+    "node-sass": "^4.13.0",
+    "react-styleguidist": "^9.2.0",
+    "react-test-renderer": "^16.11.0",
     "sass-loader": "^7.3.1",
     "storybook-react-router": "^1.0.5"
   },
